requirements.txt
streamlit==1.12.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.25.1
msrest==0.7.1
Pillow==9.0.1